feat(auth/customers): adds user verification in customer management

+++ src/use-cases/Customer/DeleteCustomerUseCase.ts
@@ -1,3 +1,4 @@
+import { ERRORS_MESSAGES } from "../../constants/Errors";
 import { CustomerRepository } from "../../repositories/CustomerRepository";
 
 export class DeleteCustomerUseCase {
@@ -6,7 +7,13 @@ export class DeleteCustomerUseCase {
     private customerRepository: CustomerRepository
   ) {}
 
-  async execute(customerUuid: string) {
+  async execute(userId: number, customerUuid: string) {
+    const customer = await this.customerRepository.getByUuid(customerUuid);
+
+    if (customer?.userId !== userId) {
+      throw new Error(ERRORS_MESSAGES.UNAUTHORIZED);
+    }
+
     await this.customerRepository.delete(customerUuid);
   }
 }
